FHIR: minor config refactor.
