+++ src/main/java/org/snomed/snowstorm/fhir/services/HapiParametersMapper.java
@@ -11,12 +11,14 @@ import org.snomed.snowstorm.fhir.domain.FHIRConcept;
 import org.snomed.snowstorm.fhir.domain.FHIRDesignation;
 import org.snomed.snowstorm.fhir.domain.FHIRProperty;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+@Service
 public class HapiParametersMapper implements FHIRConstants {
 	
 	@Autowired
